Created a BuildingController class that creates the list of buildings. (localhost:9000/buildings).
Made changes to Building class to include BuildingId which is used when finding all the doors for that building.

Made changes to the Application class to load the BuildingController class on the server. Included messages mirroring the messages for segments and locations.

TO DO:
Change the dropdown list for start and end locations to use Buildings instead of Locations. Once the Building is chosen, generate all routes between buildings (their doors) and picks the door that returns the shortest path.

Love Ari and Jim <3